+++ NAStify/ConnectionManager/Plugins/Box/CMBox.h
@@ -62,7 +62,7 @@
 #endif
 
 /* Server features */
-- (NSInteger)supportedFeaturesAtPath:(NSString *)path;
+- (long long)supportedFeaturesAtPath:(NSString *)path;
 #ifndef APP_EXTENSION
 - (NSInteger)supportedSharingFeatures;
 #endif
