+++ NAStify/ConnectionManager/Plugins/OneDrive/CMOneDrive.m
@@ -376,9 +376,9 @@
 
 #pragma mark - supported features
 
-- (NSInteger)supportedFeaturesAtPath:(NSString *)path
+- (long long)supportedFeaturesAtPath:(NSString *)path
 {
-    NSInteger features = CMSupportedFeaturesMaskFolderCreate   |
+    long long features = CMSupportedFeaturesMaskFolderCreate   |
                          CMSupportedFeaturesMaskFileDelete     |
                          CMSupportedFeaturesMaskFolderDelete   |
                          CMSupportedFeaturesMaskDeleteCancel   |
