+++ NAStify/ConnectionManager/Plugins/OwnCloud/CMOwnCloud.h
@@ -66,6 +66,6 @@
 #endif
 
 /* Server features */
-- (NSInteger)supportedFeaturesAtPath:(NSString *)path;
+- (long long)supportedFeaturesAtPath:(NSString *)path;
 
 @end
