+++ NAStify/ConnectionManager/Plugins/QNAP/CMQnap.m
@@ -2657,9 +2657,9 @@ else if (([JSON isKindOfClass:[NSDictionary class]]) && \
 
 #pragma mark - supported features
 
-- (NSInteger)supportedFeaturesAtPath:(NSString *)path
+- (long long)supportedFeaturesAtPath:(NSString *)path
 {
-    NSInteger features = CMSupportedFeaturesNone;
+    long long features = CMSupportedFeaturesNone;
     if (![path isEqualToString:@"/"])
     {
         features = CMSupportedFeaturesMaskFileDelete      |
