+++ NAStify/ConnectionManager/Plugins/WebDAV/CMWebDav.m
@@ -1042,9 +1042,9 @@ void status(void *userdata, ne_session_status status,
 #endif
 
 /* Server features */
-- (NSInteger)supportedFeaturesAtPath:(NSString *)path
+- (long long)supportedFeaturesAtPath:(NSString *)path
 {
-    NSInteger features =  CMSupportedFeaturesMaskFileDelete      |
+    long long features =  CMSupportedFeaturesMaskFileDelete      |
                           CMSupportedFeaturesMaskFolderDelete    |
                           CMSupportedFeaturesMaskFolderCreate    |
                           CMSupportedFeaturesMaskFileRename      |
